"averageRating" rounding added

+++ src/main/kotlin/com/unrec/imdb/rater/model/YearStatistics.kt
@@ -1,7 +1,9 @@
 package com.unrec.imdb.rater.model
 
+import java.math.BigDecimal
+
 data class YearStatistics(
     val year: Short,
     val count: Int,
-    val averageRating: Float
+    val averageRating: BigDecimal
 )
